Mudando tipo da interface e criando conversor para alertaDTO

+++ src/main/java/com/panicButton/panicButton/repository/AlertaRepository.java
@@ -4,4 +4,4 @@ import com.panicButton.panicButton.domain.Alerta;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AlertaRepository extends CrudRepository<Alerta, Integer> {}
+public interface AlertaRepository extends CrudRepository<Alerta, Long> {}
